src/requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.26.4
packaging==23.2
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
